src/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebSocket support
python-socketio==5.11.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# MQTT client
paho-mqtt==1.6.1

# Redis client
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
